+++ Lemonade Stand/Assets/Scripts/LemonadeStand.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 
@@ -19,14 +20,8 @@ public class LemonadeStand : MonoBehaviour
     // time in seconds to serve, MIN is 1.2f
     float serve_interval = 3f;
 
-    // tracks popularity
-    float popularity = 0f;
-
-    // factor at which popularity grows
-    float popularity_factor = 0.1f;
-
     // how many seconds is one day
-    float day_duration = 60f;
+    float day_duration = 180f;
 
     // tracks the number of days
     int day_count = 1;
@@ -54,24 +49,32 @@ public class LemonadeStand : MonoBehaviour
 
     IEnumerator SellLemonade()
     {
+        // set end day to false
+        bool end_day = false;
+
         // set day timer
         float day_timer = day_duration;
 
         // set serve timer
         float serve_timer = serve_interval;
 
-        // calculate spawn interval based on popularity
+        // set spawn delay timer
+        float spawn_delay_timer = 5f;
 
-        float base_spawn_rate = 0.2f; // TODO make based off attraction
+        // set max customers 
+        int max_customers = 5;
 
-        float growth_spawn_rate = popularity_factor * (Mathf.Pow(popularity, 2.0f) / 100);
+        // set spawn duration for spawn interval
+        float spawn_duration = 5f;
 
-        float spawn_rate = base_spawn_rate + growth_spawn_rate;
-
-        float spawn_interval = 1 / spawn_rate;
+        // set spawn interval
+        float spawn_interval = spawn_duration / max_customers;
 
         // set spawn timer
         float spawn_timer = spawn_interval;
+
+        // set end early timer
+        float end_early_timer = 5f;
 
         // tracks earnings for the day
         decimal earnings = 0m;
@@ -82,30 +85,88 @@ public class LemonadeStand : MonoBehaviour
         // tracks how many customers are waiting
         int customer_queue = 0;
 
+        // tracks customers spawned
+        int customers_spawned = 0;
+
+        float customer_patience = 5f;
+
+        List<float> patience_queue = new List<float>();
+
         // reset customer queue and served display
         customers_text.text = $"Queue: {customer_queue}";
         served_text.text = $"Served: {served}";
 
-        while (day_timer > 0f)
+        while (!end_day)
         {
+            // end day if day timer or end early timer is up
+            if (day_timer <= 0 || end_early_timer <= 0) end_day = true;
+
             // decrease day and spawn timers
             day_timer -= Time.deltaTime;
-            spawn_timer -= Time.deltaTime;
 
             // update day timer text
             day_timer_text.text = $"{day_timer}";
-
-            // spawn a customer if timer is up and reset timer
-            if (spawn_timer <= 0)
+            
+            // decrease spawn delay timer until it's complete
+            if (spawn_delay_timer > 0f)
             {
-                // add customer to queue
-                customer_queue++;
+                spawn_delay_timer -= Time.deltaTime;
+            }
 
-                // update customer display
-                customers_text.text = $"Queue: {customer_queue}";
+            // if spawn delay is done and not at max customers
+            if (spawn_delay_timer <= 0f && customers_spawned < max_customers)
+            {
+                // decrease spawn timer
+                spawn_timer -= Time.deltaTime;
 
-                // reset spawn timer
-                spawn_timer += spawn_interval;
+                // spawn a customer if timer is up and reset timer
+                if (spawn_timer <= 0f)
+                {
+                    // add customer to queue
+                    customer_queue++;
+
+                    // track their patience
+                    patience_queue.Add(customer_patience);
+
+                    // track spawn
+                    customers_spawned++;
+
+                    // update customer display
+                    customers_text.text = $"Queue: {customer_queue}";
+
+                    // reset spawn timer
+                    spawn_timer += spawn_interval;
+                }
+            }
+
+            // decrease and check customer patience
+            for (int i = patience_queue.Count - 1; i >= 0; i--)
+            {
+                // front no longer loses patience
+                if (i == 0)
+                    continue;
+
+                // decrease this patience timer
+                patience_queue[i] -= Time.deltaTime;
+
+                // if patience timer is up
+                if (patience_queue[i] <= 0f)
+                {
+                    // stop tracking their patience
+                    patience_queue.RemoveAt(i);
+
+                    // remove them from queue
+                    customer_queue--;
+                    
+                    // update customer queue text
+                    customers_text.text = $"Queue: {customer_queue}";
+                }
+            }
+
+            // if there are no more customers, run end early timer
+            if (customer_queue == 0 && customers_spawned == max_customers)
+            {
+                end_early_timer -= Time.deltaTime;
             }
 
             // if someone is in queue and you have stock, start serving
@@ -115,13 +176,16 @@ public class LemonadeStand : MonoBehaviour
             }
 
             // if serving is done
-            if (serve_timer <= 0)
+            if (serve_timer <= 0 && customer_queue > 0)
             {
                 // reset serve timer
                 serve_timer += serve_interval;
 
                 // remove them from the queue
                 customer_queue--;
+
+                // stop tracking their patience
+                patience_queue.RemoveAt(0);
 
                 // serve one stock
                 stock--;
@@ -147,6 +211,9 @@ public class LemonadeStand : MonoBehaviour
 
         // End of Day
 
+        // Raise day count
+        day_count++;
+
         // change day counter display
         day_counter_text.text = $"Day: {day_count}";
 
@@ -155,16 +222,6 @@ public class LemonadeStand : MonoBehaviour
 
         // update stock display
         stock_text.text = $"{stock}";
-
-        // calculate popularity change and update popularity
-
-        float normalization_cap = 5; // caps popularity change to +/- 5
-        float sum_scores = 0.8f * served; // TODO make based on sum of scores
-
-        // change is based on the scores normalized by number served up to a cap, is 0 if none served
-        float popularity_change = served > 0 ? (sum_scores / served) * normalization_cap : 0;
-
-        popularity += popularity_change;
     }
 
     public void BuyStock()